@@ -1,7 +1,6 @@
 {
-    "name": "cogpowered/finediff",
+    "name": "dh4/finediff",
     "description": "PHP implementation of a Fine granularity Diff engine",
-    "homepage": "https://github.com/cogpowered/FineDiff",
     "license": "MIT",
     "keywords": ["finediff", "diff", "text", "string", "opcode"],
     "authors": [
@@ -9,30 +8,40 @@
             "name": "Raymond Hill"
         },
         {
-            "name": "Rob Crowe",
-            "email": "rob@cogpowered.com"
+            "name": "Rob Crowe"
+        },
+        {
+            "name": "Hannes Van De Vreken",
+            "email": "vandevreken.hannes@gmail.com"
+        },
+        {
+            "name": "Igor Finagin",
+            "email": "igor@finag.in"
         }
     ],
     "type": "library",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.2.0"
     },
     "require-dev": {
-        "mockery/mockery": "*",
-        "phpunit/phpunit": "^4 || ^5 || ^6 || ^7"
-    },
-    "config": {
-        "optimize-autoloader": true,
-        "sort-packages": true
+        "mockery/mockery": "^1.4",
+        "phpunit/phpunit": "^9.2"
     },
     "autoload": {
         "psr-4": {
-            "CogPowered\\FineDiff\\": "src/"
+            "FineDiff\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "FineDiffTests\\": "tests/"
+            "FineDiff\\Tests\\": "tests/"
         }
+    },
+    "scripts": {
+        "tests": "phpunit"
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "sort-packages": true
     }
 }
