@@ -1,39 +1,40 @@
 <?php
 
-namespace FineDiffTests\Granularity;
+namespace FineDiff\Tests\Granularity;
 
-use FineDiffTests\TestCase;
-use CogPowered\FineDiff\Delimiters;
-use CogPowered\FineDiff\Granularity\Word;
+use FineDiff\Delimiters;
+use FineDiff\Granularity\GranularityInterface;
+use FineDiff\Granularity\Word;
+use PHPUnit\Framework\TestCase as TestCase;
 
 class WordTest extends TestCase
 {
     /**
-     * @var \CogPowered\FineDiff\Granularity\Granularity
+     * @var GranularityInterface
      */
-    protected $character;
+    protected $granularity;
 
-    protected $delimiters = array(
+    /**
+     * @var array
+     */
+    protected $delimiters = [
         Delimiters::PARAGRAPH,
         Delimiters::SENTENCE,
         Delimiters::WORD,
-    );
+    ];
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->character = new Word;
+        $this->granularity = new Word();
     }
 
     public function testExtendsAndImplements()
     {
-        $this->assertInstanceOf('CogPowered\FineDiff\Granularity\Granularity', $this->character);
-        $this->assertInstanceOf('CogPowered\FineDiff\Granularity\GranularityInterface', $this->character);
-        $this->assertInstanceOf('ArrayAccess', $this->character);
-        $this->assertInstanceOf('Countable', $this->character);
+        $this->assertInstanceOf(GranularityInterface::class, $this->granularity);
     }
 
     public function testGetDelimiters()
     {
-        $this->assertEquals($this->character->getDelimiters(), $this->delimiters);
+        $this->assertEquals($this->granularity->getDelimiters(), $this->delimiters);
     }
 }
