@@ -1,69 +1,72 @@
 <?php
 
-namespace FineDiffTests\Diff;
+namespace FineDiff\Tests\Diff;
 
-use FineDiffTests\TestCase;
+use FineDiff\Granularity\Character;
+use FineDiff\Parser\OperationCodesInterface;
+use FineDiff\Parser\ParserInterface;
+use FineDiff\Render\Html;
 use Mockery as m;
-use CogPowered\FineDiff\Diff;
+use FineDiff\Diff;
+use PHPUnit\Framework\TestCase;
 
 class DependencyInjectTest extends TestCase
 {
-    public function tearDown()
+    public function tearDown(): void
     {
         m::close();
     }
 
     public function testGetGranularity()
     {
-        $character = m::mock('CogPowered\FineDiff\Granularity\Character');
+        $character = m::mock(Character::class);
 
         $diff = new Diff($character);
         $granularity = $diff->getGranularity();
 
-        $this->assertInstanceOf('CogPowered\FineDiff\Granularity\Character', $granularity);
+        $this->assertInstanceOf(Character::class, $granularity);
     }
 
     public function testGetRenderer()
     {
-        $html = m::mock('CogPowered\FineDiff\Render\Html');
+        $html = m::mock(Html::class);
 
         $diff = new Diff(null, $html);
         $renderer = $diff->getRenderer();
 
-        $this->assertInstanceOf('CogPowered\FineDiff\Render\Html', $renderer);
+        $this->assertInstanceOf(Html::class, $renderer);
     }
 
     public function testRender()
     {
-        $operation_codes = m::mock('CogPowered\FineDiff\Parser\OperationCodes');
+        $operation_codes = m::mock(OperationCodesInterface::class);
         $operation_codes->shouldReceive('generate')->andReturn('c12');
 
-        $parser = m::mock('CogPowered\FineDiff\Parser\Parser');
+        $parser = m::mock(ParserInterface::class);
         $parser->shouldReceive('parse')->andReturn($operation_codes);
 
-        $html = m::mock('CogPowered\FineDiff\Render\Html');
+        $html = m::mock(Html::class);
         $html->shouldReceive('process')->with('hello', $operation_codes)->once();
-
 
         $diff = new Diff(null, $html, $parser);
         $diff->render('hello', 'hello2');
-        
+
         $this->assertTrue(true);
     }
 
     public function testGetParser()
     {
-        $parser = m::mock('CogPowered\FineDiff\Parser\Parser');
+        $parser = m::mock(ParserInterface::class);
 
         $diff = new Diff(null, null, $parser);
         $parser = $diff->getParser();
 
-        $this->assertInstanceOf('CogPowered\FineDiff\Parser\Parser', $parser);
+        $this->assertInstanceOf(ParserInterface::class, $parser);
     }
 
     public function testGetOperationCodes()
     {
-        $parser = m::mock('CogPowered\FineDiff\Parser\Parser');
+        $parser = m::mock(ParserInterface::class);
         $parser->shouldReceive('parse')->with('foobar', 'eggfooba')->once();
 
         $diff = new Diff(null, null, $parser);
