@@ -1,33 +1,36 @@
 <?php
 
-namespace FineDiffTests\Diff;
+namespace FineDiff\Tests\Diff;
 
-use FineDiffTests\TestCase;
+use FineDiff\Diff;
+use FineDiff\Granularity\GranularityInterface;
+use FineDiff\Parser\ParserInterface;
+use FineDiff\Render\Html;
 use Mockery as m;
-use CogPowered\FineDiff\Diff;
+use PHPUnit\Framework\TestCase;
 
 class SetTest extends TestCase
 {
     /**
-     * @var \CogPowered\FineDiff\Diff
+     * @var Diff
      */
     protected $diff;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->diff = new Diff;
+        $this->diff = new Diff();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         m::close();
     }
 
     public function testSetParser()
     {
-        $this->assertFalse( method_exists($this->diff->getParser(), 'fooBar') );
+        $this->assertFalse(method_exists($this->diff->getParser(), 'fooBar'));
 
-        $parser = m::mock('CogPowered\FineDiff\Parser\Parser');
+        $parser = m::mock(ParserInterface::class);
         $parser->shouldReceive('fooBar')->once();
 
         $this->diff->setParser($parser);
@@ -38,9 +41,9 @@ class SetTest extends TestCase
 
     public function testSetRenderer()
     {
-        $this->assertFalse( method_exists($this->diff->getRenderer(), 'fooBar') );
+        $this->assertFalse(method_exists($this->diff->getRenderer(), 'fooBar'));
 
-        $html = m::mock('CogPowered\FineDiff\Render\Html');
+        $html = m::mock(Html::class);
         $html->shouldReceive('fooBar')->once();
 
         $this->diff->setRenderer($html);
@@ -51,12 +54,12 @@ class SetTest extends TestCase
 
     public function testSetGranularity()
     {
-        $this->assertFalse( method_exists($this->diff->getGranularity(), 'fooBar') );
+        $this->assertFalse(method_exists($this->diff->getGranularity(), 'fooBar'));
 
-        $granularity = m::mock('CogPowered\FineDiff\Granularity\Word');
+        $granularity = m::mock(GranularityInterface::class);
         $granularity->shouldReceive('fooBar')->once();
 
-        $parser = m::mock('CogPowered\FineDiff\Parser\Parser');
+        $parser = m::mock(ParserInterface::class);
         $parser->shouldReceive('setGranularity')->with($granularity)->once();
         $parser->shouldReceive('getGranularity')->andReturn($granularity)->once();
 
