@@ -1,31 +1,33 @@
 <?php
 
-namespace FineDiffTests\Render\Html;
+namespace FineDiff\Tests\Render\Html;
 
-use FineDiffTests\TestCase;
+use FineDiff\Parser\OperationCodes;
+use FineDiff\Render\Html;
+use FineDiff\Render\RendererInterface;
 use Mockery as m;
-use CogPowered\FineDiff\Render\Html;
+use PHPUnit\Framework\TestCase;
 
 class ProcessTest extends TestCase
 {
     /**
-     * @var \CogPowered\FineDiff\Render\RendererInterface
+     * @var RendererInterface
      */
     protected $html;
-    
-    public function setUp()
+
+    public function setUp(): void
     {
-        $this->html = new Html;
+        $this->html = new Html();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         m::close();
     }
 
     public function testProcess()
     {
-        $operation_codes = m::mock('CogPowered\FineDiff\Parser\OperationCodes');
+        $operation_codes = m::mock(OperationCodes::class);
         $operation_codes->shouldReceive('generate')->andReturn('c5i:2c6d')->once();
 
         $html = $this->html->process('Hello worlds', $operation_codes);
